Showing filtered list correctly.
